:sparkles: Feat: 사용자 경험 개선 - 특정 날짜 누를 때 해당 날짜로 뜨게끔 설정 (#6)
