+++ app/components/calendar/ScheduleFormModal.tsx
@@ -1,5 +1,5 @@
 // components/ScheduleFormModal.tsx
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import {
   Modal,
   View,
@@ -8,7 +8,6 @@ import {
   TextInput,
   TouchableOpacity,
   Platform,
-  StyleSheet,
   TouchableWithoutFeedback,
   KeyboardAvoidingView,
   Pressable,
@@ -17,6 +16,7 @@ import { SafeAreaView } from 'react-native-safe-area-context';
 import DateTimePicker from '@react-native-community/datetimepicker';
 import dayjs from 'dayjs';
 import { Ionicons } from '@expo/vector-icons';
+import styles from './ScheduleFormModal.style';
 
 interface ScheduleFormModalProps {
   visible: boolean;
@@ -25,115 +25,6 @@ interface ScheduleFormModalProps {
   onSubmit: (schedule: { title: string; date: string; start: Date; end: Date }) => void;
 }
 
-const styles = StyleSheet.create({
-  overlay: {
-    flex: 1,
-    backgroundColor: 'rgba(0,0,0,0.3)',
-    justifyContent: 'flex-end',
-    alignItems: 'center',
-  },
-  modalBox: {
-    backgroundColor: '#fff',
-    padding: 20,
-    paddingBottom: 32,
-    borderTopLeftRadius: 16,
-    borderTopRightRadius: 16,
-    width: '100%',
-    minHeight: '50%',
-  },
-  title: {
-    fontWeight: 'bold',
-    fontSize: 16,
-    marginBottom: 12,
-  },
-  titleInput: {
-    fontSize: 16,
-    fontWeight: 'bold',
-    borderBottomWidth: 1,
-    borderBottomColor: '#ccc',
-    paddingBottom: 12,
-    marginBottom: 12,
-  },
-  input: {
-    borderWidth: 1,
-    borderColor: '#ccc',
-    padding: 10,
-    borderRadius: 6,
-    marginBottom: 12,
-    width: '100%',
-  },
-  submitButton: {
-    backgroundColor: '#FFB338',
-    paddingTop: 10,
-    paddingBottom: 10,
-    paddingLeft: 16,
-    paddingRight: 16,
-    borderRadius: 4,
-  },
-  submitText: {
-    color: '#fff',
-    textAlign: 'center',
-    fontWeight: 'bold',
-  },
-  closeButton: {
-    marginTop: 10,
-  },
-  closeText: {
-    textAlign: 'center',
-    color: '#888',
-  },
-  header: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginBottom: 20,
-  },
-  dateContainer: {
-    flexDirection: 'column',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginBottom: 10,
-  },
-  dateItem: {
-    flexDirection: 'column',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginBottom: 4,
-  },
-  dateItemTitle: {
-    fontSize: 16,
-    fontWeight: 'bold',
-    marginBottom: 12,
-  },
-  dateRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    marginBottom: 12,
-  },
-
-  dateLabel: {
-    width: 50,
-    fontSize: 14,
-    fontWeight: 'bold',
-    textAlign: 'center',
-  },
-
-  dateButton: {
-    flex: 1,
-    borderWidth: 1,
-    borderColor: '#ccc',
-    borderRadius: 8,
-    paddingVertical: 8,
-    paddingHorizontal: 12,
-    marginLeft: 8,
-  },
-
-  dateText: {
-    fontSize: 14,
-    color: '#000',
-  },
-});
-
 export default function ScheduleFormModal({
   visible,
   date,
@@ -141,11 +32,28 @@ export default function ScheduleFormModal({
   onSubmit,
 }: ScheduleFormModalProps) {
   const [title, setTitle] = useState('');
-  const [start, setStart] = useState(new Date());
-  const [end, setEnd] = useState(new Date());
+  const [start, setStart] = useState(() => {
+    const parsed = dayjs(date);
+    return parsed.isValid() ? parsed.toDate() : new Date();
+  });
+
+  const [end, setEnd] = useState(() => {
+    const parsed = dayjs(date);
+    return parsed.isValid() ? parsed.add(1, 'day').toDate() : new Date();
+  });
+
+  useEffect(() => {
+    const parsed = dayjs(date);
+    if (parsed.isValid()) {
+      setStart(parsed.toDate());
+      setEnd(parsed.add(1, 'day').toDate());
+    }
+  }, [date]);
+
   const [showStart, setShowStart] = useState(false);
   const [showEnd, setShowEnd] = useState(false);
 
+  // 제출 클릭 시
   const handleSubmit = () => {
     if (onSubmit) {
       onSubmit({ title, date, start, end });
