+++ app/types/Schedule.ts
@@ -1,0 +1,7 @@
+export interface Schedule {
+  houseId: number;
+  title: string;
+  startAt: string;
+  endAt: string;
+  memberId: number;
+}
